added songs and api protection

+++ aboutme-backend/src/modules/urls/urls.controller.ts
@@ -1,12 +1,13 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post } from '@nestjs/common';
+import { Body, Controller, Headers, Delete, Get, Param, Patch, Post, UnauthorizedException } from '@nestjs/common';
 import { UrlsService } from './urls.service';
 import { Url } from './entities';
 import { CreateUrlDto, UpdateUrlDto } from './dto';
+import { ConfigService } from '@nestjs/config';
 
 @Controller('urls')
 export class UrlsController {
-    constructor(private readonly userService: UrlsService) {};
-  
+    constructor(private readonly configService: ConfigService, private readonly userService: UrlsService) {};
+
     @Get()
     getUsers(): Promise<Url[]> {
       return this.userService.getUsers();
@@ -16,18 +17,24 @@ export class UrlsController {
       return this.userService.getUser(id);
     }
     @Post()
-    createUser(@Body() body: CreateUrlDto): Promise<Url> { //TODO add protection
-      console.log("[Urls] create - "+body.name.toString());
-      return this.userService.createUser(body);
+    createUser(@Headers('apiKey') apiKey: string, @Body() body: CreateUrlDto): Promise<Url> {
+      if(apiKey === this.configService.get("API_KEY")) {
+        console.log("[Urls] create - "+body.name.toString());
+        return this.userService.createUser(body);
+      } else throw new UnauthorizedException("Invalid apiKey");
     }
     @Patch(":id")
-    updateUser(@Param("id") id: number, @Body() body: UpdateUrlDto): Promise<Url> { //TODO add protection
-      console.log("[Urls] update - "+body.name.toString());
-      return this.userService.updateUser(id, body);
+    updateUser(@Headers('apiKey') apiKey: string, @Param("id") id: number, @Body() body: UpdateUrlDto): Promise<Url> {
+      if(apiKey === this.configService.get("API_KEY")) {
+        console.log("[Urls] update - "+body.name.toString());
+        return this.userService.updateUser(id, body);
+      } else throw new UnauthorizedException("Invalid apiKey");
     }
     @Delete(":id")
-    deleteUser(@Param("id") id: number): Promise<JSON> { //TODO add protection
-      console.log("[Urls] delete - "+id);
-      return this.userService.deleteUser(id);
+    deleteUser(@Headers('apiKey') apiKey: string, @Param("id") id: number): Promise<JSON> {
+      if(apiKey === this.configService.get("API_KEY")) {
+        console.log("[Urls] delete - "+id);
+        return this.userService.deleteUser(id);
+      } else throw new UnauthorizedException("Invalid apiKey");
     }
 }
